requirements.txt
pyyaml>=5.1,<=5.2